Change hard-coded dimension form three to two
Sometimes I want to simulate two-dimensional systems. Ideally the user should
be able to switch dimensionality based on template parameter or macro. But,
for now I create a git branch to see if what changes should be made in 2D
simulation.

This change does not pass tests because of the changes in fundamental parts.